fix(parsers): preserve runtime html parser semantics
