+++ packages/parsers/src/htmlParser.ts
@@ -13,7 +13,8 @@ import type {
 } from "./types.js";
 import { validateCompositionGsap } from "./gsapSerialize";
 import { parseCompositionVariables } from "./compositionVariables.js";
-import { parseHtmlWithHfIds, walkCompositionDescendants } from "./hfIds.js";
+import { walkCompositionDescendants } from "./hfIds.js";
+import { assignHfIds } from "./hfIdAssignment.js";
 import { parseGsapScriptAcornForWrite } from "./gsapParserAcorn.js";
 import { queryByAttr } from "./utils/cssSelector.js";
 import { removeAnimationFromScript } from "./gsapWriterAcorn.js";
@@ -178,16 +179,18 @@ function resolveResolutionFromDimensions(width: number, height: number): CanvasR
 }
 
 export function parseHtml(html: string): ParsedHtml {
-  const doc = parseHtmlWithHfIds(html);
+  const parser = new DOMParser();
+  const doc = parser.parseFromString(html, "text/html");
 
   const elements: TimelineElement[] = [];
   const keyframes: Record<string, Keyframe[]> = {};
   let idCounter = 0;
 
   const htmlEl = doc.documentElement;
-  if (!htmlEl || (!/<!doctype|<html[\s>]/i.test(html) && !doc.body.firstElementChild)) {
+  if (!htmlEl) {
     throw new CompositionHtmlParseError("parseHtml: input HTML is empty or could not be parsed");
   }
+  if (doc.body) assignHfIds(doc.body);
   const customStylesAttr = htmlEl.getAttribute("data-custom-styles");
   let customStyles: string | null = null;
   if (customStylesAttr) {
